backend/python/requirements.txt
flask
gunicorn
wapiti3
requests
beautifulsoup4
apscheduler
dnspython
humanize
sslyze
python-whois
feedparser
androguard
yara-python
pyclamd
scapy

